Basic tag clouds can now be requested.
You make a call just like Shop.internal_tags_cloud :items => 3, :groups => 2   in which Shop is the model and InternalTag is the tag class.  This returns an array of InternalTag-objects that have an extra attribute named tag_score.  The tag_score is the relative score of this item vs the other items, it must be a number between 1 and groups (including both ends of the interval) and can be used as a display-size in the visualization of the tag cloud.