Tag clouds can now be generated from a Model, whilst only counting those instances of the model that have all the given tags.
This means that Shop.tags_cloud( :tags => [InternalTag.find_by_name cheap] ) will return a tag_cloud whilst counting only those shops that have the tag 'cheap' connected to it.

To keep the speed up to par, the calculations are done in one database-query.  Do take into account that this is a rather complicated query, so it may still be usefull to cache this for heavy loads.