BUGFIX + Adding :like and :not_like syntax for tag_cloud
The bugfix concerns a wrong has_and_belongs_to_many linking from the Tag to the model the Tag is connected to.

The new syntax allows for an SQL-like LIKE syntax to remove Tags from the result of tag-clouds.  This could be used to allow a certain prefix or postfix to be in the name of the Tag.